feat(app): gate native runtime behind feature for WASM editor
Add `native-runtime` feature on sqyre-app so Trunk builds (`--no-default-features`)
skip executor, capture, and vision while the editor still uses sqyre-ui-model
for logs, runtime vars, and execution highlights. Provide WASM stubs for screen
preview and window picking, and route catalog adapters through sqyre-ports.
