+++ crates/sqyre-app/src/icon_cache.rs
@@ -3,7 +3,7 @@
 use crate::assets;
 use crate::demo_icons;
 use eframe::egui::{self, ColorImage, TextureHandle, TextureOptions, Vec2};
-use sqyre_capture::ProcessIcon;
+use crate::window_types::ProcessIcon;
 use sqyre_domain::PROGRAM_DELIMITER;
 use sqyre_persist::ProgramCatalog;
 use std::collections::HashMap;
@@ -112,11 +112,25 @@ impl IconCache {
         if self.process_missing.contains_key(&key) {
             return None;
         }
-        let Some(icon) = sqyre_capture::process_icon(process_path, window_title) else {
+        #[cfg(feature = "native-runtime")]
+        {
+            let Some(icon) = sqyre_capture::process_icon(process_path, window_title).map(|i| {
+                ProcessIcon {
+                    width: i.width,
+                    height: i.height,
+                    rgba: i.rgba,
+                }
+            }) else {
+                self.process_missing.insert(key, ());
+                return None;
+            };
+            return Some(self.insert_process_icon(ctx, &key, &icon));
+        }
+        #[cfg(not(feature = "native-runtime"))]
+        {
             self.process_missing.insert(key, ());
-            return None;
-        };
-        Some(self.insert_process_icon(ctx, &key, &icon))
+            None
+        }
     }
 
     /// OS process icon for a catalog program (via its bound `process_path`).
