+++ crates/sqyre-app/src/lib.rs
@@ -10,10 +10,12 @@ mod app_run;
 mod assets;
 mod catalog;
 mod chord_record;
+#[cfg(feature = "native-runtime")]
 mod collection_capture;
 mod data_editor;
 mod data_editor_preview;
 mod demo_icons;
+#[cfg(feature = "native-runtime")]
 mod diag;
 pub mod docs_fixture;
 mod file_dialogs;
@@ -23,13 +25,21 @@ mod icon_variants;
 mod image_view;
 mod key_record;
 mod macro_meta;
+#[cfg(feature = "native-runtime")]
 mod macro_overlay;
 mod overlay_icons;
 mod paint_ctx;
 mod pickers;
+#[cfg(feature = "native-runtime")]
 mod pixel_color;
+#[cfg(feature = "native-runtime")]
+#[path = "preview_tooltip.rs"]
+mod preview_tooltip;
+#[cfg(not(feature = "native-runtime"))]
+#[path = "preview_tooltip_stub.rs"]
 mod preview_tooltip;
 mod recorded_action;
+#[cfg(feature = "native-runtime")]
 mod recording_overlay;
 mod run_session;
 mod settings;
@@ -59,6 +69,7 @@ mod widgets;
 #[cfg(target_os = "windows")]
 mod win_focused_keys;
 mod workspace;
+mod window_types;
 
 pub use settings::SettingsUi;
 
@@ -71,13 +82,11 @@ use hotkey_record::HotkeyRecordUi;
 use icon_cache::IconCache;
 use key_record::KeyRecordUi;
 use macro_meta::MacroMetaUi;
-use macro_overlay::MacroOverlay;
 use parking_lot::Mutex;
 use preview_tooltip::PreviewTooltipCache;
-use recording_overlay::RecordingOverlay;
 use run_session::RunSession;
 use sqyre_domain::Macro;
-use sqyre_executor::{SharedActionLog, SharedHighlighter, SharedRuntimeVars};
+use sqyre_ui_model::{SharedActionLog, SharedHighlighter, SharedRuntimeVars};
 use sqyre_hotkeys::{default_hotkeys, HotkeyCallbacks, HotkeyService, ScreenClickBridge};
 use sqyre_persist::{Database, ProgramCatalog, UserSettings};
 use std::sync::Arc;
@@ -89,9 +98,10 @@ use workspace::Workspace;
 #[cfg(not(target_arch = "wasm32"))]
 pub fn run() -> eframe::Result<()> {
     let _ = sqyre_persist::initialize_directories();
+    #[cfg(feature = "native-runtime")]
     diag::install(sqyre_persist::sqyre_dir());
     sqyre_update::cleanup_stale_update();
-    #[cfg(target_os = "linux")]
+    #[cfg(all(not(target_arch = "wasm32"), feature = "native-runtime"))]
     install_x11_secondary_error_hook();
 
     let instance_lock = match single_instance::try_acquire() {
@@ -134,7 +144,7 @@ pub fn run() -> eframe::Result<()> {
 }
 
 /// Keep winit from storing X errors that originate on Sqyre's secondary Displays.
-#[cfg(all(not(target_arch = "wasm32"), target_os = "linux"))]
+#[cfg(all(not(target_arch = "wasm32"), feature = "native-runtime"))]
 fn install_x11_secondary_error_hook() {
     winit::platform::x11::register_xlib_error_hook(Box::new(|display, _event| {
         sqyre_capture::owns_secondary_x_display(display)
@@ -175,9 +185,11 @@ pub struct SqyreApp {
     /// Window was hidden because a point/search-area recording is armed.
     hidden_for_recording: bool,
     /// Outline windows for live search-area selection rect.
-    recording_overlay: RecordingOverlay,
+    #[cfg(feature = "native-runtime")]
+    recording_overlay: crate::recording_overlay::RecordingOverlay,
     /// Always-on-top floating buttons that start macros.
-    macro_overlay: MacroOverlay,
+    #[cfg(feature = "native-runtime")]
+    macro_overlay: crate::macro_overlay::MacroOverlay,
     /// Left macro-list side panel visibility.
     macro_list_open: bool,
     /// Filter text for the macro list (name / tags fuzzy match).
@@ -288,41 +300,47 @@ impl SqyreApp {
         };
 
         let platform_warning = {
-            #[cfg(all(
-                not(target_arch = "wasm32"),
-                any(target_os = "linux", target_os = "windows")
-            ))]
+            #[cfg(feature = "native-runtime")]
             {
-                // Ensure OCR data exists (downloads eng.traineddata when missing).
-                let ocr_warning = match sqyre_vision::shared_leptess() {
-                    Ok(_) => None,
-                    Err(e) => {
-                        let warning = format!("OCR unavailable: {e}");
-                        eprintln!("sqyre: {warning}");
-                        Some(warning)
+                #[cfg(all(
+                    not(target_arch = "wasm32"),
+                    any(target_os = "linux", target_os = "windows")
+                ))]
+                {
+                    // Ensure OCR data exists (downloads eng.traineddata when missing).
+                    let ocr_warning = match sqyre_vision::shared_leptess() {
+                        Ok(_) => None,
+                        Err(e) => {
+                            let warning = format!("OCR unavailable: {e}");
+                            eprintln!("sqyre: {warning}");
+                            Some(warning)
+                        }
+                    };
+                    #[cfg(target_os = "linux")]
+                    {
+                        sqyre_capture::linux_session_capture_warning()
+                            .or_else(|| {
+                                match sqyre_capture::shared_capturer() {
+                                    Ok(_) => None,
+                                    Err(e) => Some(format!("Screen capture unavailable: {e}")),
+                                }
+                            })
+                            .or(ocr_warning)
                     }
-                };
-                #[cfg(target_os = "linux")]
-                {
-                    sqyre_capture::linux_session_capture_warning()
-                        .or_else(|| {
-                            // Soft probe: if DISPLAY is set but capturer still fails, surface the error.
-                            match sqyre_capture::shared_capturer() {
-                                Ok(_) => None,
-                                Err(e) => Some(format!("Screen capture unavailable: {e}")),
-                            }
-                        })
-                        .or(ocr_warning)
+                    #[cfg(target_os = "windows")]
+                    {
+                        ocr_warning
+                    }
                 }
-                #[cfg(target_os = "windows")]
+                #[cfg(not(any(
+                    all(target_os = "linux", not(target_arch = "wasm32")),
+                    all(target_os = "windows", not(target_arch = "wasm32"))
+                )))]
                 {
-                    ocr_warning
+                    None
                 }
             }
-            #[cfg(not(any(
-                all(target_os = "linux", not(target_arch = "wasm32")),
-                all(target_os = "windows", not(target_arch = "wasm32"))
-            )))]
+            #[cfg(not(feature = "native-runtime"))]
             {
                 None
             }
@@ -364,8 +382,10 @@ impl SqyreApp {
             settings_ui,
             variables_panel: variables_panel::VariablesPanelUi::default(),
             hidden_for_recording: false,
-            recording_overlay: RecordingOverlay::new(),
-            macro_overlay: MacroOverlay::new(),
+            #[cfg(feature = "native-runtime")]
+            recording_overlay: recording_overlay::RecordingOverlay::new(),
+            #[cfg(feature = "native-runtime")]
+            macro_overlay: macro_overlay::MacroOverlay::new(),
             macro_list_open: true,
             macro_list_filter: String::new(),
             tray: tray::SystemTray::default(),
