+++ crates/sqyre-persist/src/programs/mod.rs
@@ -1,7 +1,9 @@
 //! Program catalog parsed from `db.yaml` programs section.
 
 mod encode;
+mod mutate;
 mod parse;
+mod resolve;
 mod seed_general;
 mod types;
 mod util;
@@ -14,14 +16,11 @@ pub use types::{
     ProgramPoint, ProgramSearchArea,
 };
 
-use crate::fs_name::{confined_join_or_invalid, is_safe_fs_entity_name, validate_fs_entity_name};
+use crate::fs_name::{confined_join_or_invalid, is_safe_fs_entity_name};
 use crate::{images_path, PersistError, Result};
 use encode::*;
 use parse::*;
-use serde_yaml::{Mapping, Value};
-use sqyre_domain::{resolve_scalar_int, CoordinateRef, Macro, PROGRAM_DELIMITER};
-use sqyre_ports::PortError;
-use std::collections::BTreeMap;
+use serde_yaml::Value;
 use std::path::PathBuf;
 use util::*;
 
@@ -137,592 +136,6 @@ impl ProgramCatalog {
             dir.join(format!("{mask}.png"))
         } else {
             dir.join("__invalid__.png")
-        }
-    }
-
-    pub fn lookup_point(
-        &self,
-        r: &CoordinateRef,
-        resolution_key: &str,
-    ) -> std::result::Result<&ProgramPoint, PortError> {
-        Ok(self.lookup_point_sourced(r, resolution_key)?.0)
-    }
-
-    /// Point plus the resolution bucket key that supplied it (for remapping).
-    fn lookup_point_sourced<'a>(
-        &'a self,
-        r: &CoordinateRef,
-        resolution_key: &str,
-    ) -> std::result::Result<(&'a ProgramPoint, &'a str, &'a ProgramData), PortError> {
-        if r.is_collection() {
-            return Err(PortError::invalid(format!(
-                "point lookup does not accept collection ref {r:?}"
-            )));
-        }
-        let name = r.name();
-        if name.is_empty() {
-            return Err(PortError::invalid("empty point reference"));
-        }
-        if let Some(prog) = r.program() {
-            let (pt, src) = point_from(self, prog, name, resolution_key)?;
-            let data = self
-                .programs
-                .get(prog)
-                .ok_or_else(|| PortError::not_found(format!("program {prog:?} not found")))?;
-            return Ok((pt, src, data));
-        }
-        for prog in self.programs.keys() {
-            if let Ok((pt, src)) = point_from(self, prog, name, resolution_key) {
-                let data = self.programs.get(prog).expect("program exists");
-                return Ok((pt, src, data));
-            }
-        }
-        Err(PortError::not_found(format!("point {name:?} not found")))
-    }
-
-    pub fn lookup_search_area(
-        &self,
-        r: &CoordinateRef,
-        resolution_key: &str,
-    ) -> std::result::Result<&ProgramSearchArea, PortError> {
-        Ok(self.lookup_search_area_sourced(r, resolution_key)?.0)
-    }
-
-    fn lookup_search_area_sourced<'a>(
-        &'a self,
-        r: &CoordinateRef,
-        resolution_key: &str,
-    ) -> std::result::Result<(&'a ProgramSearchArea, &'a str, &'a ProgramData), PortError> {
-        if r.is_collection() {
-            return Err(PortError::invalid(format!(
-                "search area lookup does not accept collection ref {r:?}"
-            )));
-        }
-        let name = r.name();
-        if name.is_empty() {
-            return Err(PortError::invalid("empty search area reference"));
-        }
-        if let Some(prog) = r.program() {
-            let (sa, src) = search_area_from(self, prog, name, resolution_key)?;
-            let data = self
-                .programs
-                .get(prog)
-                .ok_or_else(|| PortError::not_found(format!("program {prog:?} not found")))?;
-            return Ok((sa, src, data));
-        }
-        for prog in self.programs.keys() {
-            if let Ok((sa, src)) = search_area_from(self, prog, name, resolution_key) {
-                let data = self.programs.get(prog).expect("program exists");
-                return Ok((sa, src, data));
-            }
-        }
-        Err(PortError::not_found(format!(
-            "search area {name:?} not found"
-        )))
-    }
-
-    pub fn lookup_collection(
-        &self,
-        r: &CoordinateRef,
-    ) -> std::result::Result<&ProgramCollection, PortError> {
-        let name = r.name();
-        if name.is_empty() {
-            return Err(PortError::invalid("empty collection reference"));
-        }
-        if let Some(prog) = r.program() {
-            return collection_from(self, prog, name);
-        }
-        for prog in self.programs.keys() {
-            if let Ok(c) = collection_from(self, prog, name) {
-                return Ok(c);
-            }
-        }
-        Err(PortError::not_found(format!(
-            "collection {name:?} not found"
-        )))
-    }
-
-    pub fn resolve_point(
-        &self,
-        r: &CoordinateRef,
-        macro_: &Macro,
-    ) -> std::result::Result<(i32, i32), PortError> {
-        if r.is_collection() {
-            let (lx, ty, rx, by) = self.resolve_search_area(r, macro_)?;
-            return Ok(((lx + rx) / 2, (ty + by) / 2));
-        }
-        let key = self.resolution_key().to_string();
-        let (pt, src_key, data) = self.lookup_point_sourced(r, &key)?;
-        let x = resolve_scalar_int(&pt.x, &macro_.variables)
-            .map_err(|e| PortError::invalid(format!("point X: {e}")))?;
-        let y = resolve_scalar_int(&pt.y, &macro_.variables)
-            .map_err(|e| PortError::invalid(format!("point Y: {e}")))?;
-        self.remap_xy(x, y, src_key, data)
-    }
-
-    pub fn resolve_search_area(
-        &self,
-        r: &CoordinateRef,
-        macro_: &Macro,
-    ) -> std::result::Result<(i32, i32, i32, i32), PortError> {
-        if let Some((r1, c1, r2, c2)) = r.cell_range() {
-            return self.resolve_collection_cells(r, macro_, r1, c1, r2, c2);
-        }
-        let key = self.resolution_key().to_string();
-        let (sa, src_key, data) = self.lookup_search_area_sourced(r, &key)?;
-        let lx = resolve_scalar_int(&sa.left_x, &macro_.variables).map_err(PortError::invalid)?;
-        let ty = resolve_scalar_int(&sa.top_y, &macro_.variables).map_err(PortError::invalid)?;
-        let rx = resolve_scalar_int(&sa.right_x, &macro_.variables).map_err(PortError::invalid)?;
-        let by = resolve_scalar_int(&sa.bottom_y, &macro_.variables).map_err(PortError::invalid)?;
-        let (lx, ty) = self.remap_xy(lx, ty, src_key, data)?;
-        let (rx, by) = self.remap_xy(rx, by, src_key, data)?;
-        Ok((lx, ty, rx, by))
-    }
-
-    fn remap_xy(
-        &self,
-        x: i32,
-        y: i32,
-        src_key: &str,
-        data: &ProgramData,
-    ) -> std::result::Result<(i32, i32), PortError> {
-        let rt_key = self.resolution_key();
-        if rt_key.is_empty() {
-            return Ok((x, y));
-        }
-        let (src_w, src_h) = parse_resolution_key(src_key)?;
-        let (rt_w, rt_h) = parse_resolution_key(rt_key)?;
-        let src_scale = bucket_scale(data, src_key);
-        let rt_scale = self.runtime_scale();
-        Ok((
-            remap_coord(x, src_w, rt_w, src_scale, rt_scale),
-            remap_coord(y, src_h, rt_h, src_scale, rt_scale),
-        ))
-    }
-
-    fn resolve_collection_cells(
-        &self,
-        r: &CoordinateRef,
-        macro_: &Macro,
-        r1: i32,
-        c1: i32,
-        r2: i32,
-        c2: i32,
-    ) -> std::result::Result<(i32, i32, i32, i32), PortError> {
-        let col = self.lookup_collection(r)?;
-        if col.search_area.is_empty() {
-            return Err(PortError::invalid(format!(
-                "collection {:?} has no search area",
-                col.name
-            )));
-        }
-        let sa_ref = match r.program() {
-            Some(prog) => CoordinateRef(format!("{prog}{PROGRAM_DELIMITER}{}", col.search_area)),
-            None => CoordinateRef(col.search_area.clone()),
-        };
-        let (left_x, top_y, right_x, bottom_y) = self.resolve_search_area(&sa_ref, macro_)?;
-        cell_rect(
-            left_x, top_y, right_x, bottom_y, col.rows, col.cols, r1, c1, r2, c2,
-        )
-    }
-
-    /// `program~item` → icon PNG paths (variants + legacy).
-    pub fn variant_paths(&self, target: &str) -> Vec<PathBuf> {
-        let Some((program, item)) = split_target(target) else {
-            return Vec::new();
-        };
-        let dir = self.icons_dir(program);
-        let mut paths = Vec::new();
-        if let Ok(rd) = std::fs::read_dir(&dir) {
-            let prefix = format!("{item}{PROGRAM_DELIMITER}");
-            let legacy = format!("{item}.png");
-            for entry in rd.flatten() {
-                let name = entry.file_name();
-                let name = name.to_string_lossy();
-                if name == legacy || (name.starts_with(&prefix) && name.ends_with(".png")) {
-                    paths.push(entry.path());
-                }
-            }
-        }
-        paths.sort();
-        paths
-    }
-
-    pub fn mask_path(&self, target: &str) -> Option<PathBuf> {
-        let (program, item) = split_target(target)?;
-        let item = self.programs.get(program)?.items.get(item)?;
-        if item.mask.is_empty() {
-            return None;
-        }
-        let path = self.mask_image_path(program, &item.mask);
-        if path.is_file() {
-            Some(path)
-        } else {
-            None
-        }
-    }
-
-    pub fn item_meta(&self, target: &str) -> Option<sqyre_ports::ItemMeta> {
-        let (program, item) = split_target(target)?;
-        let item = self.programs.get(program)?.items.get(item)?;
-        Some(sqyre_ports::ItemMeta {
-            name: item.name.clone(),
-            stack_max: item.stack_max,
-            cols: item.grid_cols,
-            rows: item.grid_rows,
-        })
-    }
-
-    pub fn programs_mut(&mut self) -> &mut BTreeMap<String, ProgramData> {
-        self.bump_generation();
-        &mut self.programs
-    }
-
-    /// Encode typed catalog fields, merging `masks` / `collections` / unknown keys from `previous`.
-    pub fn to_yaml_value(&self, previous: &Value) -> Value {
-        let empty_root = Mapping::new();
-        let prev_map = match previous {
-            Value::Mapping(m) => m,
-            _ => &empty_root,
-        };
-        let empty_prog = Mapping::new();
-        let mut out = Mapping::new();
-        for (name, data) in &self.programs {
-            let prev_prog = prev_map
-                .get(Value::String(name.clone()))
-                .and_then(|v| v.as_mapping())
-                .unwrap_or(&empty_prog);
-            out.insert(Value::String(name.clone()), encode_program(data, prev_prog));
-        }
-        Value::Mapping(out)
-    }
-
-    pub fn create_program(&mut self, name: impl Into<String>) -> Result<()> {
-        let name = name.into();
-        validate_fs_entity_name(name.trim())?;
-        let name = name.trim().to_string();
-        if self.programs.contains_key(&name) {
-            return Err(PersistError::Message(format!(
-                "program {name:?} already exists"
-            )));
-        }
-        let res = self.default_resolution_key();
-        let scale = self.runtime_scale();
-        let mut data = ProgramData {
-            name: name.clone(),
-            ..Default::default()
-        };
-        if !res.is_empty() {
-            data.points.insert(res.clone(), BTreeMap::new());
-            data.search_areas.insert(res.clone(), BTreeMap::new());
-            data.coord_scales.insert(res, scale);
-        }
-        self.programs.insert(name, data);
-        self.bump_generation();
-        Ok(())
-    }
-
-    pub fn rename_program(&mut self, old: &str, new: &str) -> Result<()> {
-        let new = new.trim();
-        validate_fs_entity_name(new)?;
-        if old == new {
-            return Ok(());
-        }
-        if self.programs.contains_key(new) {
-            return Err(PersistError::Message(format!(
-                "program {new:?} already exists"
-            )));
-        }
-        let mut data = self
-            .programs
-            .remove(old)
-            .ok_or_else(|| PersistError::Message(format!("program {old:?} not found")))?;
-        data.name = new.to_string();
-        self.programs.insert(new.to_string(), data);
-        self.bump_generation();
-        Ok(())
-    }
-
-    pub fn delete_program(&mut self, name: &str) -> Result<()> {
-        if self.programs.remove(name).is_none() {
-            return Err(PersistError::Message(format!("program {name:?} not found")));
-        }
-        // Only touch the filesystem when the name cannot escape the images root.
-        if is_safe_fs_entity_name(name) {
-            let icons = self.icons_dir(name);
-            let masks = self.masks_dir(name);
-            let collections = self.collections_dir(name);
-            let _ = std::fs::remove_dir_all(icons);
-            let _ = std::fs::remove_dir_all(masks);
-            let _ = std::fs::remove_dir_all(collections);
-        }
-        self.bump_generation();
-        Ok(())
-    }
-
-    /// Bind a catalog program to a running OS window (`process_path` + `window_title`).
-    pub fn set_process_binding(
-        &mut self,
-        program: &str,
-        process_path: impl Into<String>,
-        window_title: impl Into<String>,
-    ) -> Result<()> {
-        let p = self.program_mut(program)?;
-        p.process_path = process_path.into();
-        p.window_title = window_title.into();
-        Ok(())
-    }
-
-    pub fn upsert_item(&mut self, program: &str, item: ProgramItem) -> Result<()> {
-        let key = item.name.clone();
-        upsert_named_entity(self, program, key, item, |p| &mut p.items)
-    }
-
-    pub fn rename_item(&mut self, program: &str, old: &str, new: &str) -> Result<()> {
-        let new = new.trim();
-        validate_fs_entity_name(new)?;
-        {
-            let p = self.program_mut(program)?;
-            rename_keyed_map(&mut p.items, old, new, "item", |item, n| item.name = n)?;
-        }
-        if old != new && is_safe_fs_entity_name(program) && is_safe_fs_entity_name(old) {
-            self.rename_item_icon_files(program, old, new);
-        }
-        Ok(())
-    }
-
-    /// Move `{old}.png` and `{old}~*.png` icon files to the new item name.
-    fn rename_item_icon_files(&self, program: &str, old: &str, new: &str) {
-        let dir = self.icons_dir(program);
-        let Ok(rd) = std::fs::read_dir(&dir) else {
-            return;
-        };
-        let prefix = format!("{old}{PROGRAM_DELIMITER}");
-        let legacy = format!("{old}.png");
-        for entry in rd.flatten() {
-            let name = entry.file_name();
-            let name = name.to_string_lossy();
-            let dest_name = if name.as_ref() == legacy {
-                format!("{new}.png")
-            } else if name.starts_with(&prefix) && name.ends_with(".png") {
-                format!("{new}{PROGRAM_DELIMITER}{}", &name[prefix.len()..])
-            } else {
-                continue;
-            };
-            let dest = dir.join(dest_name);
-            let _ = std::fs::rename(entry.path(), dest);
-        }
-    }
-
-    pub fn delete_item(&mut self, program: &str, name: &str) -> Result<()> {
-        delete_named_entity(self, program, name, "item", |p| &mut p.items)
-    }
-
-    pub fn upsert_point(&mut self, program: &str, point: ProgramPoint) -> Result<()> {
-        let key = point.name.clone();
-        upsert_resolution_entity(self, program, key, point, |p| &mut p.points)
-    }
-
-    pub fn rename_point(&mut self, program: &str, old: &str, new: &str) -> Result<()> {
-        rename_resolution_entity(
-            self,
-            program,
-            old,
-            new,
-            "point",
-            |p| &mut p.points,
-            |pt, n| pt.name = n,
-        )
-    }
-
-    pub fn delete_point(&mut self, program: &str, name: &str) -> Result<()> {
-        delete_resolution_entity(self, program, name, "point", |p| &mut p.points)
-    }
-
-    pub fn upsert_search_area(&mut self, program: &str, area: ProgramSearchArea) -> Result<()> {
-        let key = area.name.clone();
-        upsert_resolution_entity(self, program, key, area, |p| &mut p.search_areas)
-    }
-
-    pub fn rename_search_area(&mut self, program: &str, old: &str, new: &str) -> Result<()> {
-        rename_resolution_entity(
-            self,
-            program,
-            old,
-            new,
-            "search area",
-            |p| &mut p.search_areas,
-            |sa, n| sa.name = n,
-        )?;
-        // Propagate to collection.search_area references within this program.
-        if old != new {
-            let p = self.program_mut(program)?;
-            for col in p.collections.values_mut() {
-                if col.search_area == old {
-                    col.search_area = new.to_string();
-                }
-            }
-        }
-        Ok(())
-    }
-
-    pub fn delete_search_area(&mut self, program: &str, name: &str) -> Result<()> {
-        delete_resolution_entity(self, program, name, "search area", |p| &mut p.search_areas)?;
-        let p = self.program_mut(program)?;
-        for col in p.collections.values_mut() {
-            if col.search_area == name {
-                col.search_area.clear();
-            }
-        }
-        Ok(())
-    }
-
-    pub fn upsert_mask(&mut self, program: &str, mask: ProgramMask) -> Result<()> {
-        let key = mask.name.clone();
-        upsert_named_entity(self, program, key, mask, |p| &mut p.masks)
-    }
-
-    pub fn rename_mask(&mut self, program: &str, old: &str, new: &str) -> Result<()> {
-        let new = new.trim();
-        validate_fs_entity_name(new)?;
-        let old_path = self.mask_image_path(program, old);
-        let new_path = self.mask_image_path(program, new);
-        let p = self.program_mut(program)?;
-        rename_keyed_map(&mut p.masks, old, new, "mask", |mask, n| mask.name = n)?;
-        // Propagate to item.mask references within this program.
-        for item in p.items.values_mut() {
-            if item.mask == old {
-                item.mask = new.to_string();
-            }
-        }
-        if old != new
-            && is_safe_fs_entity_name(program)
-            && is_safe_fs_entity_name(old)
-            && old_path.is_file()
-        {
-            if let Some(parent) = new_path.parent() {
-                let _ = std::fs::create_dir_all(parent);
-            }
-            let _ = std::fs::rename(&old_path, &new_path);
-        }
-        Ok(())
-    }
-
-    pub fn delete_mask(&mut self, program: &str, name: &str) -> Result<()> {
-        let path = self.mask_image_path(program, name);
-        delete_named_entity(self, program, name, "mask", |p| &mut p.masks)?;
-        let p = self.program_mut(program)?;
-        for item in p.items.values_mut() {
-            if item.mask == name {
-                item.mask.clear();
-            }
-        }
-        if is_safe_fs_entity_name(program) && is_safe_fs_entity_name(name) {
-            let _ = std::fs::remove_file(path);
-        }
-        Ok(())
-    }
-
-    pub fn upsert_collection(
-        &mut self,
-        program: &str,
-        collection: ProgramCollection,
-    ) -> Result<()> {
-        let key = collection.name.clone();
-        upsert_named_entity(self, program, key, collection, |p| &mut p.collections)
-    }
-
-    pub fn rename_collection(&mut self, program: &str, old: &str, new: &str) -> Result<()> {
-        let new = new.trim();
-        validate_fs_entity_name(new)?;
-        let old_path = self.collection_image_path(program, old);
-        let new_path = self.collection_image_path(program, new);
-        {
-            let p = self.program_mut(program)?;
-            rename_keyed_map(&mut p.collections, old, new, "collection", |col, n| {
-                col.name = n
-            })?;
-            // Propagate to atlas member lists within this program.
-            if old != new {
-                for atlas in p.atlases.values_mut() {
-                    for member in atlas.collections.iter_mut() {
-                        if member == old {
-                            *member = new.to_string();
-                        }
-                    }
-                }
-            }
-        }
-        if old != new
-            && is_safe_fs_entity_name(program)
-            && is_safe_fs_entity_name(old)
-            && old_path.is_file()
-        {
-            if let Some(parent) = new_path.parent() {
-                let _ = std::fs::create_dir_all(parent);
-            }
-            let _ = std::fs::rename(&old_path, &new_path);
-        }
-        Ok(())
-    }
-
-    pub fn delete_collection(&mut self, program: &str, name: &str) -> Result<()> {
-        let path = self.collection_image_path(program, name);
-        delete_named_entity(self, program, name, "collection", |p| &mut p.collections)?;
-        let p = self.program_mut(program)?;
-        for atlas in p.atlases.values_mut() {
-            atlas.collections.retain(|c| c != name);
-        }
-        if is_safe_fs_entity_name(program) && is_safe_fs_entity_name(name) {
-            let _ = std::fs::remove_file(path);
-        }
-        Ok(())
-    }
-
-    pub fn upsert_atlas(&mut self, program: &str, atlas: ProgramAtlas) -> Result<()> {
-        let key = atlas.name.clone();
-        upsert_named_entity(self, program, key, atlas, |p| &mut p.atlases)
-    }
-
-    pub fn rename_atlas(&mut self, program: &str, old: &str, new: &str) -> Result<()> {
-        let new = new.trim();
-        validate_fs_entity_name(new)?;
-        let p = self.program_mut(program)?;
-        rename_keyed_map(&mut p.atlases, old, new, "atlas", |atlas, n| atlas.name = n)
-    }
-
-    pub fn delete_atlas(&mut self, program: &str, name: &str) -> Result<()> {
-        delete_named_entity(self, program, name, "atlas", |p| &mut p.atlases)
-    }
-
-    pub fn lookup_atlas(
-        &self,
-        program: &str,
-        name: &str,
-    ) -> std::result::Result<&ProgramAtlas, PortError> {
-        let p = self
-            .programs
-            .get(program)
-            .ok_or_else(|| PortError::not_found(format!("program {program:?} not found")))?;
-        p.atlases
-            .get(name)
-            .ok_or_else(|| PortError::not_found(format!("atlas {name:?} not in {program}")))
-    }
-
-    fn program_mut(&mut self, name: &str) -> Result<&mut ProgramData> {
-        if !self.programs.contains_key(name) {
-            return Err(PersistError::Message(format!("program {name:?} not found")));
-        }
-        self.bump_generation();
-        Ok(self.programs.get_mut(name).expect("program exists"))
-    }
-
-    fn default_resolution_key(&self) -> String {
-        let key = self.resolution_key();
-        if key.is_empty() {
-            "1920x1080".into()
-        } else {
-            key.to_string()
         }
     }
 }
